fix(api-types): inline InlineSourcingSchema so apps/web webpack can resolve it
CI's build-nextjs failed with:
  Module not found: Can't resolve './routes/tablebase/sourcing-schema.js'

api-types.ts is bundled by apps/web (via the @wiki-server/api-types path
alias), and webpack cannot follow the deep relative `.js` import to a
file living at apps/wiki-server/src/routes/tablebase/.

Move the canonical InlineSourcingSchema definition into api-types.ts
itself and have routes/tablebase/sourcing-schema.ts re-export from
there. Single source of truth; route imports unchanged.
